rectified sidebar css and overall pading
